+++ newfrontend/src/app/app.component.html
@@ -2,7 +2,7 @@
     <h1>RIESGOS 2.0: Uncertainty analyis</h1>
 </header>
 
-<!-- <app-login></app-login> -->
+<app-login></app-login>
 
 <main>
     <div class="content">
