+++ newfrontend/src/app/services/backend/backend_api/models/bbox-input.ts
