+++ newfrontend/src/app/services/backend/backend_api/models/literal-input.ts
