+++ newfrontend/src/app/services/backend/backend_api/models/product-type.ts
