+++ newfrontend/src/app/services/backend/backend_api/strict-http-response.ts
