+++ newfrontend/src/app/services/backend/db/db.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from "@angular/core";
 import { catchError, defaultIfEmpty, forkJoin, map, Observable, of, switchMap, tap } from "rxjs";
-import { ComplexOutput, Job, Order, Process, Product, ProductType, UserSelfInformation } from "src/app/backend_api/models";
-import { ApiService } from "src/app/backend_api/services";
-import { UserOrder } from "../backend/backend.service";
+import { ComplexOutput, Job, Order, Process, Product, ProductType, UserSelfInformation } from "src/app/services/backend/backend_api/models";
+import { ApiService } from "src/app/services/backend/backend_api/services";
+import { UserOrder } from "../backend.service";
 
 
 export interface ProductInfo {
