+++ newfrontend/src/app/services/backend/logs/logs.service.ts
